[front] feat: add template compare to pptx inspect
